third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/offset-rotate-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'offset-rotate' to the 'auto' keyword: auto
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'offset-rotate' to the 'reverse' keyword: reverse
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
Harness: the test ran to completion.
